# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: |
    SubNation Digital Subscriptions Platform API.

    ## Authentication (round-93 / 93-A8 F-1 — contract truth)

    Two session mechanisms, both documented as security schemes and
    attached per operation family:

    - **User routes** (auth, orders, wallet, cart, coupons/validate,
      loyalty, referrals, support, notifications): an `auth_token`
      httpOnly cookie (primary) OR an `Authorization: Bearer <token>`
      header (fallback).
    - **Admin routes** (/admin/*, /coupons/admin*, the copilot family):
      an `admin_token` httpOnly cookie (8h, issued by POST
      /admin/login) OR its Bearer equivalent. POST /admin/login itself
      is the credential endpoint and needs no prior session.

    ## CSRF gate (applies to every state-changing request)

    Every POST/PUT/PATCH/DELETE under /api — regardless of cookie or
    Bearer authentication — must carry an Origin (or Referer) header
    matching the deployment's allow-list (CSRF_ALLOWED_ORIGINS /
    APP_ORIGINS / APP_URL); otherwise the request is rejected with
    403 FORBIDDEN before the route runs. Requests with neither Origin
    nor Referer are also rejected. Exemptions:
    /api/auth/firebase/session, /api/auth/firebase/refresh, /api/cwv,
    /api/webhook/*, and /health. Non-browser clients must therefore
    send a browser-like Origin header on mutations.

    ## Deployment-conditional behavior

...
servers:
  - url: /api
...
security:
  - cookieAuth: []
    bearerAuth: []
...
paths:
...
  /auth/whatsapp/start:
    post:
      operationId: startWhatsappOtp
      tags: [auth]
      summary: Send a WhatsApp OTP to the given phone (registration purpose)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: Libyan or international phone; normalized server-side.
                referral_code:
                  type: string
                  maxLength: 16
                  description: Optional referral code applied on first verify.
      responses:
        "200":
          description: OTP queued — never the code itself, only its expiry.
          content:
            application/json:
              schema:
                type: object
                required: [success, expires_at]
                properties:
                  success:
                    type: boolean
                  expires_at:
                    type: string
                    description: ISO timestamp when the code expires (5 minutes).
        "400":
          description: Missing/invalid phone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Cooldown (60s resend) / hourly or attempt caps — retry later.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: >-
            WhatsApp channel not ready: details.reason = whatsapp_settling
            (freshly linked channel warming up — details.retry_after_sec
            present + Retry-After header) or store_failed (delivery store
            failure, 30s retry hint).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - type: object
                    properties:
                      details:
                        type: object
                        properties:
                          reason:
                            type: string
                            enum: [whatsapp_settling, store_failed]
                          retry_after_sec:
                            type: integer
          headers:
            Retry-After:
              description: Seconds until the channel is expected to accept sends.
              schema:
                type: integer
  /auth/whatsapp/verify:
    post:
      operationId: verifyWhatsappOtp
      tags: [auth]
      summary: Verify the WhatsApp OTP — mints the session (cookie) on success
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                code:
                  type: string
                  description: 6-digit code as typed (Arabic-Indic digits accepted).
                referral_code:
                  type: string
                  maxLength: 16
      responses:
        "200":
          description: Session minted (httpOnly auth_token cookie set).
          content:
            application/json:
              schema:
                type: object
                required: [token, is_new_user]
                properties:
                  token:
                    type: string
                    description: The "__cookie_session__" sentinel — not a credential.
                  is_new_user:
                    type: boolean
        "400":
          description: Missing fields / malformed code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong code (5 attempts per code, consume-on-success)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    # Round-93 (93-A8 F-1): the spec previously documented zero security
    # definitions while the middleware reality is cookie-first with a
    # Bearer fallback on user routes, and a separate admin cookie (8h)
    # plus Bearer on admin routes. Both alternatives are expressed per
    # operation via `security: [{ schemeA: [], schemeB: [] }]` (same
    # object = OR).
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: |
        User session — httpOnly cookie set by the auth routes (Google /
        Telegram / WhatsApp-OTP sign-in). SameSite=None in production so
        the Vercel-hosted SPA can call the API cross-site. Primary
        mechanism for browser clients.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User-session Bearer fallback — `Authorization: Bearer <token>`
        accepted wherever cookieAuth is (requireUser checks the cookie
        first, then the header).
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        Admin session — separate httpOnly cookie (8h) issued by POST
        /admin/login (and verify-2fa when TOTP is enabled).
    adminBearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin-session Bearer alternative — the token returned in the
        POST /admin/login JSON body, accepted wherever adminCookieAuth
        is (requireAdmin accepts either).
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"
        details:
          type: object
          additionalProperties: true
          description: |
            Optional structured context — zod `issues` on validation
            failures, `{field}` hints on flash-sale validation errors,
            `{skipped_duplicates}` on inventory uploads.
      required: [error, code]

    # Machine-readable error codes — mirrors the shared ErrorCode enum
    # in @workspace/error-codes (shared/error-codes/src/index.ts), the
    # single source both backend and frontend re-export. The copilot
    # preview/execution family (round-92 B3/F-05) closes the gap where
    # backend/src/routes/admin/copilot/* returned codes outside the enum.
    ErrorCode:
      type: string
      enum:
        - INVALID_DATA
        - INVALID_PHONE
        - INVALID_PASSWORD_LENGTH
        - INVALID_PASSWORD_WEAK
        - INVALID_OTP
        - INVALID_CREDENTIAL
        - UNAUTHORIZED
        - INVALID_TOKEN
        - SESSION_EXPIRED
        - ACCOUNT_LOCKED
        - ACCOUNT_NOT_FOUND
        - PHONE_ALREADY_REGISTERED
        - FEATURE_DISABLED
        - FORBIDDEN
        - INSUFFICIENT_PERMISSIONS
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - OUT_OF_STOCK
        - PRODUCT_UNAVAILABLE
        - INSUFFICIENT_BALANCE
        - INVALID_AMOUNT
        - TOPUP_LIMIT_EXCEEDED
        - ORDER_NOT_FOUND
        - ORDER_ALREADY_COMPLETED
        - ORDER_CANNOT_CANCEL
        - GOOGLE_TOKEN_INVALID
        - GOOGLE_VERIFICATION_FAILED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
        - RATE_LIMITED
        - IDEMPOTENCY_IN_FLIGHT
        - IDEMPOTENCY_KEY_REUSE
        - COPILOT_INVALID_INPUT
        - COPILOT_LLM_ERROR
        - COPILOT_NO_ADMIN_SESSION
        - COPILOT_BAD_METHOD
        - COPILOT_PREVIEW_NOT_FOUND
        - COPILOT_PREVIEW_CONSUMED
        - COPILOT_PREVIEW_EXPIRED
        - COPILOT_HANDOFF_REQUIRED
        - COPILOT_HIGH_RISK_DISABLED
        - COPILOT_STALE_RECORD
        - COPILOT_UNEXPECTED_STATE
        - COPILOT_EXECUTE_FAILED
        - COPILOT_NOT_HIGH_RISK
        - COPILOT_FIRST_CONFIRM_MISSING
        - COPILOT_COOLDOWN_NOT_ELAPSED
        - COPILOT_LLM_UNAVAILABLE
        - COPILOT_OUT_OF_SCOPE
        - COPILOT_INVALID_FLAGS
        # Round-93 (93-A8 F-8/F-15): two emitted codes that previously
        # lived outside the whole code contract — the copilot rate
        # limiter's 429 body and the CSRF gate's fail-closed
        # misconfiguration branch. The enum stays exhaustive over what
        # the backend actually emits.
        - COPILOT_RATE_LIMITED
        - CSRF_CONFIG
        # Round-94 (A5-05): the last two emitted-but-unenumerated codes —
        # the copilot phase-gate 503 (requireCopilotPhase) and the
        # outbound secret-scan 502 abort (ask/draft routes).
        - COPILOT_PHASE_DISABLED
        - COPILOT_SECRET_LEAK

    # 409 body returned by the idempotency middleware
    # (backend/src/middlewares/idempotency.ts): the same Idempotency-Key
    # re-sent with a different body (IDEMPOTENCY_KEY_REUSE) or the same
    # key while the original request is still executing
    # (IDEMPOTENCY_IN_FLIGHT). `message` is a legacy alias kept for older
    # clients — `error` is the canonical field.
...